feat: override detail page
