+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_checkout/__tests__/applePayExpress.test.js
@@ -8,7 +8,7 @@ const {
   handleApplePayResponse,
   callPaymentFromComponent,
   formatCustomerObject,
-} = require('../../applePayExpress');
+} = require('../../adyen_express/applepay/applePayExpress');
 
 
 beforeEach(() => {
