+++ src/cartridges/app_adyen_SFRA/cartridge/client/default/js/adyen_express/applepay/applePayExpress.js
@@ -1,0 +1,3 @@
+const applePayExpressModule = require('./common');
+
+applePayExpressModule.init();
