--- src/cartridges/app_adyen_SFRA/cartridge/templates/default/product/components/addToCartProduct.isml
@@ -1,27 +1,0 @@
-<isset name="AdyenHelper" value="${require('*/cartridge/adyen/utils/adyenHelper')}" scope="pdict"/>
-<isset name="AdyenConfigs" value="${require('*/cartridge/adyen/utils/adyenConfigs')}" scope="page" />
-<script src="${pdict.AdyenHelper.getCheckoutUrl()}" type="text/javascript"></script>
-<link rel="stylesheet" type="text/css" href="${pdict.AdyenHelper.getCheckoutCSS()}"/>
-<link rel="stylesheet" href="${URLUtils.staticURL('/css/adyenCss.css')}" />
-<script type="text/javascript">
-        window.environment = "${AdyenConfigs.getAdyenEnvironment()}";
-        window.clientKey = "${AdyenConfigs.getAdyenClientKey()}",
-        window.locale = "${request.locale}";
-        window.getPaymentMethodsURL = "${URLUtils.https('Adyen-GetExpressPdpPaymentMethods')}";
-        window.shippingMethodsUrl = "${URLUtils.https('Adyen-ShippingMethods')}";
-</script>
-<script src="${URLUtils.httpsStatic('/js/applePayExpress.js')}" type="text/javascript"></script>
-<div class="row cart-and-ipay">
-    <div class="col-sm-12">
-        <input type="hidden" class="add-to-cart-url" value="${pdict.addToCartUrl}">
-        <button class="add-to-cart btn btn-primary"
-                data-toggle="modal" data-target="#chooseBonusProductModal"
-                data-pid="${product.id}"
-                ${!product.readyToOrder || !product.available ? "disabled" : ""}>
-            <i class="fa fa-shopping-bag"></i>
-            ${Resource.msg('button.addtocart', 'common', null)}
-        </button>
-        <div data-method="applepay" class="expressComponent applepay" style="padding:0"></div>
-        <isinclude template="product/components/addToCartButtonExtension"/>
-    </div>
-</div>
